refactor: convert plant registration to server action
